+++ Lagrange.Core/Event/EventArg/GroupInvitationRequestEvent.cs
@@ -2,7 +2,7 @@ namespace Lagrange.Core.Event.EventArg;
 
 public class GroupInvitationRequestEvent : EventBase
 {
-    internal GroupInvitationRequestEvent(uint groupUin, uint targetUin, uint invitorUin)
+    internal GroupInvitationRequestEvent(ulong groupUin, ulong targetUin, ulong invitorUin)
     {
         GroupUin = groupUin;
         TargetUin = targetUin;
@@ -10,9 +10,9 @@ public class GroupInvitationRequestEvent : EventBase
         EventMessage = $"[{nameof(GroupInvitationRequestEvent)}] {TargetUin} from {InvitorUin} at {GroupUin}";
     }
     
-    public uint GroupUin { get; }
+    public ulong GroupUin { get; }
     
-    public uint TargetUin { get; }
+    public ulong TargetUin { get; }
     
-    public uint InvitorUin { get; }
+    public ulong InvitorUin { get; }
 }
