+++ Lagrange.Core/Event/EventArg/GroupJoinRequestEvent.cs
@@ -2,14 +2,14 @@ namespace Lagrange.Core.Event.EventArg;
 
 public class GroupJoinRequestEvent : EventBase
 {
-    internal GroupJoinRequestEvent(uint groupUin, uint targetUin)
+    internal GroupJoinRequestEvent(ulong groupUin, ulong targetUin)
     {
         TargetUin = targetUin;
         GroupUin = groupUin;
         EventMessage = $"[{nameof(GroupJoinRequestEvent)}] {TargetUin} at {GroupUin}";
     }
     
-    public uint TargetUin { get; }
+    public ulong TargetUin { get; }
     
-    public uint GroupUin { get; }
+    public ulong GroupUin { get; }
 }
