+++ Lagrange.Core/Event/EventArg/GroupRecallPokeEvent.cs
@@ -2,13 +2,13 @@ namespace Lagrange.Core.Event.EventArg;
 
 public class GroupRecallPokeEvent : EventBase
 {
-    public uint GroupUin { get; }
+    public ulong GroupUin { get; }
 
     public uint OperatorUin { get; }
 
     public ulong TipsSeqId { get; set; }
 
-    public GroupRecallPokeEvent(uint groupUin, uint operatorUin, ulong tipsSeqId)
+    public GroupRecallPokeEvent(ulong groupUin, uint operatorUin, ulong tipsSeqId)
     {
         GroupUin = groupUin;
         OperatorUin = operatorUin;
