+++ Lagrange.Core/Internal/Event/Notify/GroupSysInviteEvent.cs
@@ -2,15 +2,15 @@ namespace Lagrange.Core.Internal.Event.Notify;
 
 internal class GroupSysInviteEvent : ProtocolEvent
 {
-    public uint GroupUin { get; set; }
+    public ulong GroupUin { get; set; }
     
     public string InvitorUid { get; set; }
 
-    private GroupSysInviteEvent(uint groupUin, string invitorUid) : base(0)
+    private GroupSysInviteEvent(ulong groupUin, string invitorUid) : base(0)
     {
         GroupUin = groupUin;
         InvitorUid = invitorUid;
     }
     
-    public static GroupSysInviteEvent Result(uint groupUin, string invitorUid) => new(groupUin, invitorUid);
+    public static GroupSysInviteEvent Result(ulong groupUin, string invitorUid) => new(groupUin, invitorUid);
 }
