+++ Lagrange.Core/Internal/Event/Notify/GroupSysRequestInvitationEvent.cs
@@ -1,20 +1,22 @@
+using Lagrange.Core.Internal.Event;
+
 namespace Lagrange.Core.Internal.Event.Notify;
 
 internal class GroupSysRequestInvitationEvent : ProtocolEvent
 {
-    public uint GroupUin { get; }
+    public ulong GroupUin { get; }
     
     public string TargetUid { get; }
     
     public string InvitorUid { get; }
 
-    private GroupSysRequestInvitationEvent(uint groupUin, string targetUid, string invitorUid) : base(0)
+    private GroupSysRequestInvitationEvent(ulong groupUin, string targetUid, string invitorUid) : base(0)
     {
         GroupUin = groupUin;
         TargetUid = targetUid;
         InvitorUid = invitorUid;
     }
 
-    public static GroupSysRequestInvitationEvent Result(uint groupUin, string targetUid, string invitorUid) 
+    public static GroupSysRequestInvitationEvent Result(ulong groupUin, string targetUid, string invitorUid) 
         => new(groupUin, targetUid, invitorUid);
 }
