+++ Lagrange.Core/Internal/Packets/Message/Notify/GroupInvite.cs
@@ -7,13 +7,13 @@ namespace Lagrange.Core.Internal.Packets.Message.Notify;
 [ProtoContract]
 internal class GroupInvite
 {
-    [ProtoMember(1)] public uint GroupUin { get; set; }
+    [ProtoMember(1)] public ulong GroupUin { get; set; }
     
-    [ProtoMember(2)] public uint Field2 { get; set; } // 1
+    [ProtoMember(2)] public ulong Field2 { get; set; } // 1
     
-    [ProtoMember(3)] public uint Field3 { get; set; } // 4
+    [ProtoMember(3)] public ulong Field3 { get; set; } // 4
     
-    [ProtoMember(4)] public uint Field4 { get; set; } // 0
+    [ProtoMember(4)] public ulong Field4 { get; set; } // 0
     
     [ProtoMember(5)] public string InvitorUid { get; set; }
